arreglo de diseños de vistas simulaciones
