+++ TEST-aluvion-frontend/components/SubmitButton.vue
@@ -21,13 +21,13 @@
 <style>
   .submit-button {
     background: #025cfa;
-    padding: 1rem 1.5rem 1rem 1.5rem;
+    padding: 1rem 1.25rem 1rem 1.25rem;
     font-size: 1rem;
     color: #fff;
     justify-content: center;
     text-transform: uppercase;
     font-weight: bold;
-    border-radius: 10px;
+    border-radius: 30px;
   }
   .submit-button:hover {
         background: white;
